Move list, set, frozenset and tuple tests to tests/types/
Also split the corresponding cases out of the `test_default_validators`
parameterization and replace `conlist()`/`conset()`/`confrozenset()`
usages with the explicit annotated forms.
