+++ tests/test_types.py
@@ -149,509 +149,6 @@ def test_strict_raw_type():
         Model(v=b'fo')
 
 
-def test_constrained_list_good():
-    class ConListModelMax(BaseModel):
-        v: conlist(int) = []
-
-    m = ConListModelMax(v=[1, 2, 3])
-    assert m.v == [1, 2, 3]
-
-
-def test_constrained_list_default():
-    class ConListModelMax(BaseModel):
-        v: conlist(int) = []
-
-    m = ConListModelMax()
-    assert m.v == []
-
-
-def test_constrained_list_too_long():
-    class ConListModelMax(BaseModel):
-        v: conlist(int, max_length=10) = []
-
-    with pytest.raises(ValidationError) as exc_info:
-        ConListModelMax(v=list(str(i) for i in range(11)))
-    # insert_assert(exc_info.value.errors(include_url=False))
-    assert exc_info.value.errors(include_url=False) == [
-        {
-            'type': 'too_long',
-            'loc': ('v',),
-            'msg': 'List should have at most 10 items after validation, not 11',
-            'input': ['0', '1', '2', '3', '4', '5', '6', '7', '8', '9', '10'],
-            'ctx': {'field_type': 'List', 'max_length': 10, 'actual_length': 11},
-        }
-    ]
-
-
-def test_constrained_list_too_short():
-    class ConListModelMin(BaseModel):
-        v: conlist(int, min_length=1)
-
-    with pytest.raises(ValidationError) as exc_info:
-        ConListModelMin(v=[])
-    # insert_assert(exc_info.value.errors(include_url=False))
-    assert exc_info.value.errors(include_url=False) == [
-        {
-            'type': 'too_short',
-            'loc': ('v',),
-            'msg': 'List should have at least 1 item after validation, not 0',
-            'input': [],
-            'ctx': {'field_type': 'List', 'min_length': 1, 'actual_length': 0},
-        }
-    ]
-
-
-def test_constrained_list_optional():
-    class Model(BaseModel):
-        req: conlist(str, min_length=1) | None
-        opt: conlist(str, min_length=1) | None = None
-
-    assert Model(req=None).model_dump() == {'req': None, 'opt': None}
-    assert Model(req=None, opt=None).model_dump() == {'req': None, 'opt': None}
-
-    with pytest.raises(ValidationError) as exc_info:
-        Model(req=[], opt=[])
-    # insert_assert(exc_info.value.errors(include_url=False))
-    assert exc_info.value.errors(include_url=False) == [
-        {
-            'type': 'too_short',
-            'loc': ('req',),
-            'msg': 'List should have at least 1 item after validation, not 0',
-            'input': [],
-            'ctx': {'field_type': 'List', 'min_length': 1, 'actual_length': 0},
-        },
-        {
-            'type': 'too_short',
-            'loc': ('opt',),
-            'msg': 'List should have at least 1 item after validation, not 0',
-            'input': [],
-            'ctx': {'field_type': 'List', 'min_length': 1, 'actual_length': 0},
-        },
-    ]
-
-    assert Model(req=['a'], opt=['a']).model_dump() == {'req': ['a'], 'opt': ['a']}
-
-
-def test_constrained_list_constraints():
-    class ConListModelBoth(BaseModel):
-        v: conlist(int, min_length=7, max_length=11)
-
-    m = ConListModelBoth(v=list(range(7)))
-    assert m.v == list(range(7))
-
-    m = ConListModelBoth(v=list(range(11)))
-    assert m.v == list(range(11))
-
-    with pytest.raises(ValidationError) as exc_info:
-        ConListModelBoth(v=list(range(6)))
-    # insert_assert(exc_info.value.errors(include_url=False))
-    assert exc_info.value.errors(include_url=False) == [
-        {
-            'type': 'too_short',
-            'loc': ('v',),
-            'msg': 'List should have at least 7 items after validation, not 6',
-            'input': [0, 1, 2, 3, 4, 5],
-            'ctx': {'field_type': 'List', 'min_length': 7, 'actual_length': 6},
-        }
-    ]
-
-    with pytest.raises(ValidationError) as exc_info:
-        ConListModelBoth(v=list(range(12)))
-    # insert_assert(exc_info.value.errors(include_url=False))
-    assert exc_info.value.errors(include_url=False) == [
-        {
-            'type': 'too_long',
-            'loc': ('v',),
-            'msg': 'List should have at most 11 items after validation, not 12',
-            'input': [0, 1, 2, 3, 4, 5, 6, 7, 8, 9, 10, 11],
-            'ctx': {'field_type': 'List', 'max_length': 11, 'actual_length': 12},
-        }
-    ]
-
-    with pytest.raises(ValidationError) as exc_info:
-        ConListModelBoth(v=1)
-    # insert_assert(exc_info.value.errors(include_url=False))
-    assert exc_info.value.errors(include_url=False) == [
-        {'type': 'list_type', 'loc': ('v',), 'msg': 'Input should be a valid list', 'input': 1}
-    ]
-
-
-def test_constrained_list_item_type_fails():
-    class ConListModel(BaseModel):
-        v: conlist(int) = []
-
-    with pytest.raises(ValidationError) as exc_info:
-        ConListModel(v=['a', 'b', 'c'])
-    # insert_assert(exc_info.value.errors(include_url=False))
-    assert exc_info.value.errors(include_url=False) == [
-        {
-            'type': 'int_parsing',
-            'loc': ('v', 0),
-            'msg': 'Input should be a valid integer, unable to parse string as an integer',
-            'input': 'a',
-        },
-        {
-            'type': 'int_parsing',
-            'loc': ('v', 1),
-            'msg': 'Input should be a valid integer, unable to parse string as an integer',
-            'input': 'b',
-        },
-        {
-            'type': 'int_parsing',
-            'loc': ('v', 2),
-            'msg': 'Input should be a valid integer, unable to parse string as an integer',
-            'input': 'c',
-        },
-    ]
-
-
-def test_conlist():
-    class Model(BaseModel):
-        foo: list[int] = Field(min_length=2, max_length=4)
-        bar: conlist(str, min_length=1, max_length=4) = None
-
-    assert Model(foo=[1, 2], bar=['spoon']).model_dump() == {'foo': [1, 2], 'bar': ['spoon']}
-
-    msg = r'List should have at least 2 items after validation, not 1 \[type=too_short,'
-    with pytest.raises(ValidationError, match=msg):
-        Model(foo=[1])
-
-    msg = r'List should have at most 4 items after validation, not 5 \[type=too_long,'
-    with pytest.raises(ValidationError, match=msg):
-        Model(foo=list(range(5)))
-
-    with pytest.raises(ValidationError) as exc_info:
-        Model(foo=[1, 'x', 'y'])
-    # insert_assert(exc_info.value.errors(include_url=False))
-    assert exc_info.value.errors(include_url=False) == [
-        {
-            'type': 'int_parsing',
-            'loc': ('foo', 1),
-            'msg': 'Input should be a valid integer, unable to parse string as an integer',
-            'input': 'x',
-        },
-        {
-            'type': 'int_parsing',
-            'loc': ('foo', 2),
-            'msg': 'Input should be a valid integer, unable to parse string as an integer',
-            'input': 'y',
-        },
-    ]
-
-    with pytest.raises(ValidationError) as exc_info:
-        Model(foo=1)
-    # insert_assert(exc_info.value.errors(include_url=False))
-    assert exc_info.value.errors(include_url=False) == [
-        {'type': 'list_type', 'loc': ('foo',), 'msg': 'Input should be a valid list', 'input': 1}
-    ]
-
-
-def test_conlist_wrong_type_default():
-    """It should not validate default value by default"""
-
-    class Model(BaseModel):
-        v: conlist(int) = 'a'
-
-    m = Model()
-    assert m.v == 'a'
-
-
-def test_constrained_set_good():
-    class Model(BaseModel):
-        v: conset(int) = []
-
-    m = Model(v=[1, 2, 3])
-    assert m.v == {1, 2, 3}
-
-
-def test_constrained_set_default():
-    class Model(BaseModel):
-        v: conset(int) = set()
-
-    m = Model()
-    assert m.v == set()
-
-
-def test_constrained_set_default_invalid():
-    class Model(BaseModel):
-        v: conset(int) = 'not valid, not validated'
-
-    m = Model()
-    assert m.v == 'not valid, not validated'
-
-
-def test_constrained_set_too_long():
-    class ConSetModelMax(BaseModel):
-        v: conset(int, max_length=10) = []
-
-    with pytest.raises(ValidationError) as exc_info:
-        ConSetModelMax(v={str(i) for i in range(11)})
-    # insert_assert(exc_info.value.errors(include_url=False))
-    assert exc_info.value.errors(include_url=False) == [
-        {
-            'type': 'too_long',
-            'loc': ('v',),
-            'msg': 'Set should have at most 10 items after validation, not more',
-            'input': {'4', '3', '10', '9', '5', '6', '1', '8', '0', '7', '2'},
-            'ctx': {'field_type': 'Set', 'max_length': 10, 'actual_length': None},
-        }
-    ]
-
-
-def test_constrained_set_too_short():
-    class ConSetModelMin(BaseModel):
-        v: conset(int, min_length=1)
-
-    with pytest.raises(ValidationError) as exc_info:
-        ConSetModelMin(v=[])
-    # insert_assert(exc_info.value.errors(include_url=False))
-    assert exc_info.value.errors(include_url=False) == [
-        {
-            'type': 'too_short',
-            'loc': ('v',),
-            'msg': 'Set should have at least 1 item after validation, not 0',
-            'input': [],
-            'ctx': {'field_type': 'Set', 'min_length': 1, 'actual_length': 0},
-        }
-    ]
-
-
-def test_constrained_set_optional():
-    class Model(BaseModel):
-        req: conset(str, min_length=1) | None
-        opt: conset(str, min_length=1) | None = None
-
-    assert Model(req=None).model_dump() == {'req': None, 'opt': None}
-    assert Model(req=None, opt=None).model_dump() == {'req': None, 'opt': None}
-
-    with pytest.raises(ValidationError) as exc_info:
-        Model(req=set(), opt=set())
-    # insert_assert(exc_info.value.errors(include_url=False))
-    assert exc_info.value.errors(include_url=False) == [
-        {
-            'type': 'too_short',
-            'loc': ('req',),
-            'msg': 'Set should have at least 1 item after validation, not 0',
-            'input': set(),
-            'ctx': {'field_type': 'Set', 'min_length': 1, 'actual_length': 0},
-        },
-        {
-            'type': 'too_short',
-            'loc': ('opt',),
-            'msg': 'Set should have at least 1 item after validation, not 0',
-            'input': set(),
-            'ctx': {'field_type': 'Set', 'min_length': 1, 'actual_length': 0},
-        },
-    ]
-
-    assert Model(req={'a'}, opt={'a'}).model_dump() == {'req': {'a'}, 'opt': {'a'}}
-
-
-def test_constrained_set_constraints():
-    class ConSetModelBoth(BaseModel):
-        v: conset(int, min_length=7, max_length=11)
-
-    m = ConSetModelBoth(v=set(range(7)))
-    assert m.v == set(range(7))
-
-    m = ConSetModelBoth(v=set(range(11)))
-    assert m.v == set(range(11))
-
-    with pytest.raises(ValidationError) as exc_info:
-        ConSetModelBoth(v=set(range(6)))
-    # insert_assert(exc_info.value.errors(include_url=False))
-    assert exc_info.value.errors(include_url=False) == [
-        {
-            'type': 'too_short',
-            'loc': ('v',),
-            'msg': 'Set should have at least 7 items after validation, not 6',
-            'input': {0, 1, 2, 3, 4, 5},
-            'ctx': {'field_type': 'Set', 'min_length': 7, 'actual_length': 6},
-        }
-    ]
-
-    with pytest.raises(ValidationError) as exc_info:
-        ConSetModelBoth(v=set(range(12)))
-    # insert_assert(exc_info.value.errors(include_url=False))
-    assert exc_info.value.errors(include_url=False) == [
-        {
-            'type': 'too_long',
-            'loc': ('v',),
-            'msg': 'Set should have at most 11 items after validation, not more',
-            'input': {0, 8, 1, 9, 2, 10, 3, 7, 11, 4, 6, 5},
-            'ctx': {'field_type': 'Set', 'max_length': 11, 'actual_length': None},
-        }
-    ]
-
-    with pytest.raises(ValidationError) as exc_info:
-        ConSetModelBoth(v=1)
-    # insert_assert(exc_info.value.errors(include_url=False))
-    assert exc_info.value.errors(include_url=False) == [
-        {'type': 'set_type', 'loc': ('v',), 'msg': 'Input should be a valid set', 'input': 1}
-    ]
-
-
-def test_constrained_set_item_type_fails():
-    class ConSetModel(BaseModel):
-        v: conset(int) = []
-
-    with pytest.raises(ValidationError) as exc_info:
-        ConSetModel(v=['a', 'b', 'c'])
-    # insert_assert(exc_info.value.errors(include_url=False))
-    assert exc_info.value.errors(include_url=False) == [
-        {
-            'type': 'int_parsing',
-            'loc': ('v', 0),
-            'msg': 'Input should be a valid integer, unable to parse string as an integer',
-            'input': 'a',
-        },
-        {
-            'type': 'int_parsing',
-            'loc': ('v', 1),
-            'msg': 'Input should be a valid integer, unable to parse string as an integer',
-            'input': 'b',
-        },
-        {
-            'type': 'int_parsing',
-            'loc': ('v', 2),
-            'msg': 'Input should be a valid integer, unable to parse string as an integer',
-            'input': 'c',
-        },
-    ]
-
-
-def test_conset():
-    class Model(BaseModel):
-        foo: set[int] = Field(min_length=2, max_length=4)
-        bar: conset(str, min_length=1, max_length=4) = None
-
-    assert Model(foo=[1, 2], bar=['spoon']).model_dump() == {'foo': {1, 2}, 'bar': {'spoon'}}
-
-    assert Model(foo=[1, 1, 1, 2, 2], bar=['spoon']).model_dump() == {'foo': {1, 2}, 'bar': {'spoon'}}
-
-    with pytest.raises(ValidationError, match='Set should have at least 2 items after validation, not 1'):
-        Model(foo=[1])
-
-    with pytest.raises(ValidationError, match='Set should have at most 4 items after validation, not more'):
-        Model(foo=list(range(5)))
-
-    with pytest.raises(ValidationError) as exc_info:
-        Model(foo=[1, 'x', 'y'])
-    # insert_assert(exc_info.value.errors(include_url=False))
-    assert exc_info.value.errors(include_url=False) == [
-        {
-            'type': 'int_parsing',
-            'loc': ('foo', 1),
-            'msg': 'Input should be a valid integer, unable to parse string as an integer',
-            'input': 'x',
-        },
-        {
-            'type': 'int_parsing',
-            'loc': ('foo', 2),
-            'msg': 'Input should be a valid integer, unable to parse string as an integer',
-            'input': 'y',
-        },
-    ]
-
-    with pytest.raises(ValidationError) as exc_info:
-        Model(foo=1)
-    # insert_assert(exc_info.value.errors(include_url=False))
-    assert exc_info.value.errors(include_url=False) == [
-        {'type': 'set_type', 'loc': ('foo',), 'msg': 'Input should be a valid set', 'input': 1}
-    ]
-
-
-def test_conset_not_required():
-    class Model(BaseModel):
-        foo: set[int] | None = None
-
-    assert Model(foo=None).foo is None
-    assert Model().foo is None
-
-
-def test_confrozenset():
-    class Model(BaseModel):
-        foo: frozenset[int] = Field(min_length=2, max_length=4)
-        bar: confrozenset(str, min_length=1, max_length=4) = None
-
-    m = Model(foo=[1, 2], bar=['spoon'])
-    assert m.model_dump() == {'foo': {1, 2}, 'bar': {'spoon'}}
-    assert isinstance(m.foo, frozenset)
-    assert isinstance(m.bar, frozenset)
-
-    assert Model(foo=[1, 1, 1, 2, 2], bar=['spoon']).model_dump() == {'foo': {1, 2}, 'bar': {'spoon'}}
-
-    with pytest.raises(ValidationError, match='Frozenset should have at least 2 items after validation, not 1'):
-        Model(foo=[1])
-
-    with pytest.raises(ValidationError, match='Frozenset should have at most 4 items after validation, not more'):
-        Model(foo=list(range(5)))
-
-    with pytest.raises(ValidationError) as exc_info:
-        Model(foo=[1, 'x', 'y'])
-    # insert_assert(exc_info.value.errors(include_url=False))
-    assert exc_info.value.errors(include_url=False) == [
-        {
-            'type': 'int_parsing',
-            'loc': ('foo', 1),
-            'msg': 'Input should be a valid integer, unable to parse string as an integer',
-            'input': 'x',
-        },
-        {
-            'type': 'int_parsing',
-            'loc': ('foo', 2),
-            'msg': 'Input should be a valid integer, unable to parse string as an integer',
-            'input': 'y',
-        },
-    ]
-
-    with pytest.raises(ValidationError) as exc_info:
-        Model(foo=1)
-    # insert_assert(exc_info.value.errors(include_url=False))
-    assert exc_info.value.errors(include_url=False) == [
-        {'type': 'frozen_set_type', 'loc': ('foo',), 'msg': 'Input should be a valid frozenset', 'input': 1}
-    ]
-
-
-def test_confrozenset_not_required():
-    class Model(BaseModel):
-        foo: frozenset[int] | None = None
-
-    assert Model(foo=None).foo is None
-    assert Model().foo is None
-
-
-def test_constrained_frozenset_optional():
-    class Model(BaseModel):
-        req: confrozenset(str, min_length=1) | None
-        opt: confrozenset(str, min_length=1) | None = None
-
-    assert Model(req=None).model_dump() == {'req': None, 'opt': None}
-    assert Model(req=None, opt=None).model_dump() == {'req': None, 'opt': None}
-
-    with pytest.raises(ValidationError) as exc_info:
-        Model(req=frozenset(), opt=frozenset())
-    # insert_assert(exc_info.value.errors(include_url=False))
-    assert exc_info.value.errors(include_url=False) == [
-        {
-            'type': 'too_short',
-            'loc': ('req',),
-            'msg': 'Frozenset should have at least 1 item after validation, not 0',
-            'input': frozenset(),
-            'ctx': {'field_type': 'Frozenset', 'min_length': 1, 'actual_length': 0},
-        },
-        {
-            'type': 'too_short',
-            'loc': ('opt',),
-            'msg': 'Frozenset should have at least 1 item after validation, not 0',
-            'input': frozenset(),
-            'ctx': {'field_type': 'Frozenset', 'min_length': 1, 'actual_length': 0},
-        },
-    ]
-
-    assert Model(req={'a'}, opt={'a'}).model_dump() == {'req': {'a'}, 'opt': {'a'}}
-
-
 @pytest.mark.parametrize(
     'annotation',
     [
@@ -1080,10 +577,6 @@ def check_model_fixture():
         datetime_check: datetime = datetime(2017, 5, 5, 10, 10, 10)
         time_check: time = time(10, 10, 10)
         timedelta_check: timedelta = timedelta(days=1)
-        list_check: list[str] = ['1', '2']
-        tuple_check: tuple[str, ...] = ('1', '2')
-        set_check: set[str] = {'1', '2'}
-        frozenset_check: frozenset[str] = frozenset(['1', '2'])
 
     return CheckModel
 
@@ -1251,34 +744,6 @@ class BoolCastable:
         pytest.param(
             'timedelta_check', b'1 10:10', ValidationError, id='timedelta_check-1 10:10_bytes-ValidationError'
         ),
-        ('list_check', ['1', '2'], ['1', '2']),
-        ('list_check', ('1', '2'), ['1', '2']),
-        ('list_check', {'1': 1, '2': 2}.keys(), ['1', '2']),
-        ('list_check', {'1': '1', '2': '2'}.values(), ['1', '2']),
-        ('list_check', {'1', '2'}, dirty_equals.IsOneOf(['1', '2'], ['2', '1'])),
-        ('list_check', frozenset(['1', '2']), dirty_equals.IsOneOf(['1', '2'], ['2', '1'])),
-        ('list_check', {'1': 1, '2': 2}, ValidationError),
-        ('tuple_check', ('1', '2'), ('1', '2')),
-        ('tuple_check', ['1', '2'], ('1', '2')),
-        ('tuple_check', {'1': 1, '2': 2}.keys(), ('1', '2')),
-        ('tuple_check', {'1': '1', '2': '2'}.values(), ('1', '2')),
-        ('tuple_check', {'1', '2'}, dirty_equals.IsOneOf(('1', '2'), ('2', '1'))),
-        ('tuple_check', frozenset(['1', '2']), dirty_equals.IsOneOf(('1', '2'), ('2', '1'))),
-        ('tuple_check', {'1': 1, '2': 2}, ValidationError),
-        ('set_check', {'1', '2'}, {'1', '2'}),
-        ('set_check', ['1', '2', '1', '2'], {'1', '2'}),
-        ('set_check', ('1', '2', '1', '2'), {'1', '2'}),
-        ('set_check', frozenset(['1', '2']), {'1', '2'}),
-        ('set_check', {'1': 1, '2': 2}.keys(), {'1', '2'}),
-        ('set_check', {'1': '1', '2': '2'}.values(), {'1', '2'}),
-        ('set_check', {'1': 1, '2': 2}, ValidationError),
-        ('frozenset_check', frozenset(['1', '2']), frozenset(['1', '2'])),
-        ('frozenset_check', ['1', '2', '1', '2'], frozenset(['1', '2'])),
-        ('frozenset_check', ('1', '2', '1', '2'), frozenset(['1', '2'])),
-        ('frozenset_check', {'1', '2'}, frozenset(['1', '2'])),
-        ('frozenset_check', {'1': 1, '2': 2}.keys(), frozenset(['1', '2'])),
-        ('frozenset_check', {'1': '1', '2': '2'}.values(), frozenset(['1', '2'])),
-        ('frozenset_check', {'1': 1, '2': 2}, ValidationError),
     ],
 )
 def test_default_validators(field, value, result, CheckModel):
@@ -2266,126 +1731,6 @@ def test_sequence_fails(cls, value, errors):
 
 def test_sequence_strict():
     assert TypeAdapter(Sequence[int]).validate_python((), strict=True) == ()
-
-
-def test_list_strict() -> None:
-    class LaxModel(BaseModel):
-        v: list[int]
-
-        model_config = ConfigDict(strict=False)
-
-    class StrictModel(BaseModel):
-        v: list[int]
-
-        model_config = ConfigDict(strict=True)
-
-    assert LaxModel(v=(1, 2)).v == [1, 2]
-    assert LaxModel(v=('1', 2)).v == [1, 2]
-    # Tuple should be rejected
-    with pytest.raises(ValidationError) as exc_info:
-        StrictModel(v=(1, 2))
-    assert exc_info.value.errors(include_url=False) == [
-        {'type': 'list_type', 'loc': ('v',), 'msg': 'Input should be a valid list', 'input': (1, 2)}
-    ]
-    # Strict in each list item
-    with pytest.raises(ValidationError) as exc_info:
-        StrictModel(v=['1', 2])
-    assert exc_info.value.errors(include_url=False) == [
-        {'type': 'int_type', 'loc': ('v', 0), 'msg': 'Input should be a valid integer', 'input': '1'}
-    ]
-
-
-def test_set_strict() -> None:
-    class LaxModel(BaseModel):
-        v: set[int]
-
-        model_config = ConfigDict(strict=False)
-
-    class StrictModel(BaseModel):
-        v: set[int]
-
-        model_config = ConfigDict(strict=True)
-
-    assert LaxModel(v=(1, 2)).v == {1, 2}
-    assert LaxModel(v=('1', 2)).v == {1, 2}
-    # Tuple should be rejected
-    with pytest.raises(ValidationError) as exc_info:
-        StrictModel(v=(1, 2))
-    assert exc_info.value.errors(include_url=False) == [
-        {
-            'type': 'set_type',
-            'loc': ('v',),
-            'msg': 'Input should be a valid set',
-            'input': (1, 2),
-        }
-    ]
-    # Strict in each set item
-    with pytest.raises(ValidationError) as exc_info:
-        StrictModel(v={'1', 2})
-    err_info = exc_info.value.errors(include_url=False)
-    # Sets are not ordered
-    del err_info[0]['loc']
-    assert err_info == [{'type': 'int_type', 'msg': 'Input should be a valid integer', 'input': '1'}]
-
-
-def test_frozenset_strict() -> None:
-    class LaxModel(BaseModel):
-        v: frozenset[int]
-
-        model_config = ConfigDict(strict=False)
-
-    class StrictModel(BaseModel):
-        v: frozenset[int]
-
-        model_config = ConfigDict(strict=True)
-
-    assert LaxModel(v=(1, 2)).v == frozenset((1, 2))
-    assert LaxModel(v=('1', 2)).v == frozenset((1, 2))
-    # Tuple should be rejected
-    with pytest.raises(ValidationError) as exc_info:
-        StrictModel(v=(1, 2))
-    assert exc_info.value.errors(include_url=False) == [
-        {
-            'type': 'frozen_set_type',
-            'loc': ('v',),
-            'msg': 'Input should be a valid frozenset',
-            'input': (1, 2),
-        }
-    ]
-    # Strict in each set item
-    with pytest.raises(ValidationError) as exc_info:
-        StrictModel(v=frozenset(('1', 2)))
-    err_info = exc_info.value.errors(include_url=False)
-    # Sets are not ordered
-    del err_info[0]['loc']
-    assert err_info == [{'type': 'int_type', 'msg': 'Input should be a valid integer', 'input': '1'}]
-
-
-def test_tuple_strict() -> None:
-    class LaxModel(BaseModel):
-        v: tuple[int, int]
-
-        model_config = ConfigDict(strict=False)
-
-    class StrictModel(BaseModel):
-        v: tuple[int, int]
-
-        model_config = ConfigDict(strict=True)
-
-    assert LaxModel(v=[1, 2]).v == (1, 2)
-    assert LaxModel(v=['1', 2]).v == (1, 2)
-    # List should be rejected
-    with pytest.raises(ValidationError) as exc_info:
-        StrictModel(v=[1, 2])
-    assert exc_info.value.errors(include_url=False) == [
-        {'type': 'tuple_type', 'loc': ('v',), 'msg': 'Input should be a valid tuple', 'input': [1, 2]}
-    ]
-    # Strict in each list item
-    with pytest.raises(ValidationError) as exc_info:
-        StrictModel(v=('1', 2))
-    assert exc_info.value.errors(include_url=False) == [
-        {'type': 'int_type', 'loc': ('v', 0), 'msg': 'Input should be a valid integer', 'input': '1'}
-    ]
 
 
 def test_int_validation():
@@ -4591,50 +3936,6 @@ def test_literal_multiple():
             'ctx': {'expected': "'a' or 'b'"},
         }
     ]
-
-
-def test_typing_mutable_set():
-    s1 = TypeAdapter(set[int]).core_schema
-    s1.pop('metadata', None)
-    s2 = TypeAdapter(typing.MutableSet[int]).core_schema
-    s2.pop('metadata', None)
-    assert s1 == s2
-
-
-def test_frozenset_field():
-    class FrozenSetModel(BaseModel):
-        set: frozenset[int]
-
-    test_set = frozenset({1, 2, 3})
-    object_under_test = FrozenSetModel(set=test_set)
-
-    assert object_under_test.set == test_set
-
-
-@pytest.mark.parametrize(
-    'value,result',
-    [
-        ([1, 2, 3], frozenset([1, 2, 3])),
-        ({1, 2, 3}, frozenset([1, 2, 3])),
-        ((1, 2, 3), frozenset([1, 2, 3])),
-        (deque([1, 2, 3]), frozenset([1, 2, 3])),
-    ],
-)
-def test_frozenset_field_conversion(value, result):
-    class FrozenSetModel(BaseModel):
-        set: frozenset[int]
-
-    object_under_test = FrozenSetModel(set=value)
-
-    assert object_under_test.set == result
-
-
-def test_frozenset_field_not_convertible():
-    class FrozenSetModel(BaseModel):
-        set: frozenset[int]
-
-    with pytest.raises(ValidationError, match=r'frozenset'):
-        FrozenSetModel(set=42)
 
 
 @pytest.mark.parametrize(
